pyproject.toml
[project]
name = "prompt-engineering-for-generative-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cohere>=5.20.5",
    "fastapi>=0.129.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
]
